Moved config file to according package
