+++ src/main/java/org/wahlzeit_revisited/main/ModelMain.java
@@ -24,7 +24,7 @@ package org.wahlzeit_revisited.main;
 import org.wahlzeit_revisited.database.DatabaseConnection;
 import org.wahlzeit_revisited.database.SessionManager;
 import org.wahlzeit_revisited.utils.SysLog;
-import org.wahlzeit_revisited.utils.WahlzeitConfig;
+import org.wahlzeit_revisited.config.WahlzeitConfig;
 
 import java.io.BufferedReader;
 import java.io.File;
